+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v2.4.0] - 2022-02-21
+### Added
+- A new command `lak list accounts` that allows printing account values and
+percentages by accounts or by account types.
+### Fixed
+- The spinner chars were not showing properly on MS Windows 11. Changed to
+a simpler spinner.
+
 ## [v2.3.0] - 2022-01-25
 ### Added
 - A new module `lakshmi.performance` that adds ability to checkpoint